+++ addon/components/standard-template-card.hbs
@@ -1,14 +1,16 @@
-{{#if this.hasApplicableTemplates}}
-  <AuDropdown @dropdownTitle="Voeg sjabloon in">
+<TemplateProvider @controller={{@controller}} as |provider|>
+  {{#if provider.hasApplicableTemplates}}
+    <AuDropdown @dropdownTitle="Voeg sjabloon in">
 
-    {{#unless this.busy}}
-      {{#each this.applicableTemplates as |template|}}
-        <AuButton @skin="tertiary" role="menuItem" {{on "click" (fn this.insert template)}}>
-          <AuIcon @icon="edit" @alignment="left" />
-          {{template.title}}
-        </AuButton>
-      {{/each}}
-    {{/unless}}
+      {{#unless provider.busy}}
+        {{#each provider.applicableTemplates as |template|}}
+          <AuButton @skin="tertiary" role="menuItem" {{on "click" (fn this.insert template)}}>
+            <AuIcon @icon="edit" @alignment="left" />
+            {{template.title}}
+          </AuButton>
+        {{/each}}
+      {{/unless}}
 
-  </AuDropdown>
-{{/if}}
+    </AuDropdown>
+  {{/if}}
+</TemplateProvider>
