+++ addon/components/suggested-templates-card.hbs
@@ -1,37 +1,27 @@
-{{#if this.active}}
-<div style="z-index: 999;position: relative">
-  <div class="say-card__header">
-    <h3 class="say-card__subtitle">
-      U kan hier deze sjablonen invoegen
-    </h3>
-  </div>
-  <div class="say-card__content">
-    <ul class="say-link-list link-list--bordered">
-      {{#each this.templates.loadTemplates as |template|}}
-        <li class="say-link-list__item link-list__item--sidenotes">
-          <button type="button" class="say-link-list__link" {{on "click" (fn this.insert template)}}>
-            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" class="say-icon say-icon--large"
-                 aria-hidden="true">
-              <path
-                d="M14.4,8.35714846 L0,8.35714846 L0,10.7857161 L14.4,10.7857161 L14.4,8.35714846 Z M14.4,3.5 L0,3.5 L0,5.92856769 L14.4,5.92856769 L14.4,3.5 Z M19.2,13.2142838 L19.2,8.35714846 L16.8,8.35714846 L16.8,13.2142838 L12,13.2142838 L12,15.6428646 L16.8,15.6428646 L16.8,20.5 L19.2,20.5 L19.2,15.6428646 L24,15.6428646 L24,13.2142838 L19.2,13.2142838 Z M0,15.6428646 L9.6,15.6428646 L9.6,13.2142838 L0,13.2142838 L0,15.6428646 Z"></path>
-            </svg>
-            {{template.title}}
-          </button>
-        </li>
-      {{/each}}
-    </ul>
-  </div>
-  <div class="say-card__footer say-u-text-right">
-    <button type="button" {{on "click" this.closeHints}} role="button" class="say-button say-button--tertiary">
-      <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" class="say-icon say-icon--left say-icon--large">
-        <g class="cross">
-          <path fill-rule="evenodd"
-                d="M12 13.414l-5.293 5.293-1.414-1.414L10.586 12 5.293 6.707l1.414-1.414L12 10.586l5.293-5.293 1.414 1.414L13.414 12l5.293 5.293-1.414 1.414L12 13.414z"
-                class="Vector" clip-rule="evenodd" />
-        </g>
-      </svg>
-      Venster sluiten
-    </button>
-  </div>
-</div>
-{{/if}}
+<TemplateProvider @controller={{@controller}} as |provider|>
+  {{#if provider.hasApplicableTemplates}}
+    <div style="z-index: 999;position: relative">
+      <div class="say-card__header">
+        <h3 class="say-card__subtitle">
+          U kan hier deze sjablonen invoegen
+        </h3>
+      </div>
+      <div class="say-card__content">
+        <ul class="say-link-list link-list--bordered">
+          {{#each provider.applicableTemplates as |template|}}
+            <li class="say-link-list__item link-list__item--sidenotes">
+              <button type="button" class="say-link-list__link" {{on "click" (fn this.insert template)}}>
+                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" class="say-icon say-icon--large"
+                     aria-hidden="true">
+                  <path
+                    d="M14.4,8.35714846 L0,8.35714846 L0,10.7857161 L14.4,10.7857161 L14.4,8.35714846 Z M14.4,3.5 L0,3.5 L0,5.92856769 L14.4,5.92856769 L14.4,3.5 Z M19.2,13.2142838 L19.2,8.35714846 L16.8,8.35714846 L16.8,13.2142838 L12,13.2142838 L12,15.6428646 L16.8,15.6428646 L16.8,20.5 L19.2,20.5 L19.2,15.6428646 L24,15.6428646 L24,13.2142838 L19.2,13.2142838 Z M0,15.6428646 L9.6,15.6428646 L9.6,13.2142838 L0,13.2142838 L0,15.6428646 Z"></path>
+                </svg>
+                {{template.title}}
+              </button>
+            </li>
+          {{/each}}
+        </ul>
+      </div>
+    </div>
+  {{/if}}
+</TemplateProvider>
