+++ addon/components/template-provider.hbs
@@ -1,0 +1,5 @@
+{{yield (hash
+          applicableTemplates=this.applicableTemplates
+          hasApplicableTemplates=this.hasApplicableTemplates
+          busy=this.busy
+        )}}
