--- app/components/editor-plugins/suggested-templates-card.js
@@ -1,1 +1,0 @@
-export { default } from '@lblod/ember-rdfa-editor-standard-template-plugin/components/editor-plugins/suggested-templates-card';
