+++ app/components/suggested-templates-card.js
@@ -1,0 +1,1 @@
+export { default } from '@lblod/ember-rdfa-editor-standard-template-plugin/components/suggested-templates-card';
